refactor : 의존성을 각 모듈에 맞게 분리
