+++ clog-api/build.gradle.kts
@@ -3,8 +3,17 @@ dependencies {
     implementation(project(":clog-domain"))
     implementation(project(":clog-infrastructure"))
 
-    implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.5")
+    implementation("org.springframework.boot:spring-boot-starter-validation")
+    implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.5.0")
+    implementation("org.springframework.boot:spring-boot-starter-security")
+    implementation("org.springframework.boot:spring-boot-starter-oauth2-client")
+    implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
+    implementation("com.auth0:java-jwt:4.4.0")
+    implementation("com.auth0:jwks-rsa:0.20.0")
+
+    // @EnableJpaRepositorie 해결하면 삭제해주세요
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
 }
 
 tasks {
@@ -15,5 +24,3 @@ tasks {
         enabled = true
     }
 }
-
-
