+++ clog-infrastructure/build.gradle.kts
@@ -2,12 +2,14 @@ dependencies {
     implementation(project(":clog-global-utils"))
     implementation(project(":clog-domain"))
 
-    implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.10.0")
 
     implementation("com.linecorp.kotlin-jdsl:jpql-dsl:3.5.4")
     implementation("com.linecorp.kotlin-jdsl:jpql-render:3.5.4")
+
+    implementation("com.h2database:h2")
+    runtimeOnly("com.h2database:h2")
 
     runtimeOnly("com.mysql:mysql-connector-j")
 }
